Split the code into smaller chunks
