+++ src/performers/text.js
@@ -1,0 +1,6 @@
+import { text as toText } from '../utils/formatter'
+
+export default (text) => {
+	return Promise.resolve(text)
+		.then(toText)
+}
